// src/redux/data/data.types.ts
export enum DataActionTypes {
  SET_USER_DATA = 'SET_USER_DATA',
  SET_FAT_DATA = 'SET_FAT_DATA',
  SET_WEIGHT_DATA = 'SET_WEIGHT_DATA',
  SET_FINISH_DATE = 'SET_FINISH_DATE',
  SET_DAILY_WEIGHT = 'SET_DAILY_WEIGHT',
  CLEAR_ACTUAL_GOAL = 'CLEAR_ACTUAL_GOAL',
  CLEAR_ACTUAL_GOAL_SAVE_WEIGHTS = 'CLEAR_ACTUAL_GOAL_SAVE_WEIGHTS',
  CLEAR_FINISH_DATE_ONLY = 'CLEAR_FINISH_DATE_ONLY',
  SET_FORMULA = 'SET_FORMULA'
}